# openapi/dist/aep-v1-m1.bundle.yaml
openapi: 3.1.0
info:
  title: Agent Enterprise Protocol M1
  version: 1.0.0-m1
  description: M1 contract for model discovery, authorization, and gateway connection.
  license:
    name: Apache 2.0
    identifier: Apache-2.0
servers:
  - url: http://localhost:8080
    description: Local M1 deployment
  - url: https://enterprise.example.com
    description: Protected deployment
security:
  - bearerAuth: []
...
paths:
...
  /aep/v1/admin/control-events:
    get:
      tags:
        - Admin Control Events
      operationId: listControlEvents
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/ControlEventType'
        - name: scopeType
          in: query
          schema:
            $ref: '#/components/schemas/ControlEventScopeType'
        - name: scopeId
          in: query
          schema:
            type: string
        - name: createdAfter
          in: query
          schema:
            type: string
            format: date-time
        - name: createdBefore
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Control events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminControlEventPage'
    post:
      tags:
        - Admin Control Events
      operationId: createControlEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateControlEventRequest'
      responses:
        '201':
          description: Control event published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminControlEvent'
        '400':
          $ref: '#/components/responses/Problem-2'
...
  /aep/v1/admin/control-events/{eventId}/cancel:
    post:
      tags:
        - Admin Control Events
      operationId: cancelControlEvent
      description: Cancels only deliveries that have not reached received.
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Event after cancellation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminControlEvent'
        '404':
          $ref: '#/components/responses/Problem-2'
        '409':
          $ref: '#/components/responses/Problem-2'
...
  /aep/v1/admin/events:
    get:
      tags:
        - Admin Events
      operationId: searchEvents
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit-2'
        - name: userId
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: resourceType
          in: query
          schema:
            type: string
        - name: resourceId
          in: query
          schema:
            type: string
        - name: result
          in: query
          schema:
            $ref: '#/components/schemas/EventResult'
        - name: occurredAfter
          in: query
          schema:
            type: string
            format: date-time
        - name: occurredBefore
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Event search result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventSearchResult'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Problem-2:
      type: object
      required:
        - type
        - title
        - status
        - code
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        code:
          type: string
        requestId:
          type: string
...
    ControlEventType:
      type: string
      enum:
        - skill.manifest.changed
        - plugin.manifest.changed
        - credential.assignments.changed
        - model.catalog.changed
    GlobalScope:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          const: global
      additionalProperties: false
    TeamScope:
      type: object
      required:
        - type
        - id
      properties:
        type:
          type: string
          const: team
        id:
          type: string
      additionalProperties: false
    UserScope:
      type: object
      required:
        - type
        - id
      properties:
        type:
          type: string
          const: user
        id:
          type: string
      additionalProperties: false
    ControlEventScope:
      oneOf:
        - $ref: '#/components/schemas/GlobalScope'
        - $ref: '#/components/schemas/TeamScope'
        - $ref: '#/components/schemas/UserScope'
      discriminator:
        propertyName: type
    ControlEventResource:
      type: object
      required:
        - type
        - id
        - revision
      properties:
        type:
          type: string
          enum:
            - skill
            - plugin
            - credential
            - model
        id:
          type: string
        revision:
          type: string
    ControlTask:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum:
            - skill.reconcile
            - plugin.reconcile
            - credential.reconcile
            - model.reconcile
...
    EventResource:
      type: object
      required:
        - type
        - id
      properties:
        type:
          type: string
        id:
          type: string
    EventResult:
      type: string
      enum:
        - success
        - failure
        - info
    UserEvent:
      type: object
      required:
        - eventId
        - type
        - occurredAt
        - result
      properties:
        eventId:
          type: string
        type:
          type: string
        occurredAt:
          type: string
          format: date-time
        resource:
          $ref: '#/components/schemas/EventResource'
        result:
          $ref: '#/components/schemas/EventResult'
        metadata:
          type: object
          additionalProperties: true
...
    ControlEventScopeType:
      type: string
      enum:
        - global
        - team
        - user
    CreateControlEventRequest:
      type: object
      required:
        - type
        - scope
        - task
        - expiresAt
      properties:
        type:
          $ref: '#/components/schemas/ControlEventType'
        scope:
          $ref: '#/components/schemas/ControlEventScope'
        resource:
          $ref: '#/components/schemas/ControlEventResource'
        task:
          $ref: '#/components/schemas/ControlTask'
        expiresAt:
          type: string
          format: date-time
        supersedesKey:
          type:
            - string
            - 'null'
          maxLength: 500
    DeliverySummary:
      type: object
      required:
        - pending
        - received
        - running
        - succeeded
        - failed
        - expired
        - superseded
      properties:
        pending:
          type: integer
          minimum: 0
        received:
          type: integer
          minimum: 0
        running:
          type: integer
          minimum: 0
        succeeded:
          type: integer
          minimum: 0
        failed:
          type: integer
          minimum: 0
        expired:
          type: integer
          minimum: 0
        superseded:
          type: integer
          minimum: 0
    AdminControlEvent:
      allOf:
        - $ref: '#/components/schemas/CreateControlEventRequest'
        - type: object
          required:
            - eventId
            - state
            - createdAt
            - createdBy
            - deliverySummary
          properties:
            eventId:
              type: string
            state:
              type: string
              enum:
                - active
                - cancelled
                - expired
                - superseded
            createdAt:
              type: string
              format: date-time
            createdBy:
              type: string
            deliverySummary:
              $ref: '#/components/schemas/DeliverySummary'
    AdminControlEventPage:
      type: object
      required:
        - items
        - nextCursor
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminControlEvent'
        nextCursor:
          type:
            - string
            - 'null'
...
    StoredEvent:
      allOf:
        - $ref: '#/components/schemas/UserEvent'
        - type: object
          required:
            - receivedAt
            - userId
            - sessionId
          properties:
            receivedAt:
              type: string
              format: date-time
            userId:
              type: string
            sessionId:
              type: string
    EventSearchResult:
      type: object
      required:
        - items
        - nextCursor
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/StoredEvent'
        nextCursor:
          type:
            - string
            - 'null'
...
  responses:
...
    Problem-2:
      description: AEP Problem Details
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem-2'
...
  parameters:
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
    Cursor:
      name: cursor
      in: query
      schema:
        type:
          - string
          - 'null'
    Limit-2:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    EventId:
      name: eventId
      in: path
      required: true
      schema:
        type: string
...